removes int64 from docstring (#3208)

+++ pyvista/core/pointset.py
@@ -781,8 +781,7 @@ class PolyData(_vtk.vtkPolyData, _PointSet, PolyDataFilters):
         >>> polygon = pv.Rectangle()
         >>> extruded = polygon.extrude((0, 0, 1))
         >>> extruded.strips
-        array([4, 0, 1, 4, 5, 4, 1, 2, 5, 6, 4, 2, 3, 6, 7, 4, 3, 0, 7, 4],
-              dtype=int64)
+        array([4, 0, 1, 4, 5, 4, 1, 2, 5, 6, 4, 2, 3, 6, 7, 4, 3, 0, 7, 4])
         """
         return _vtk.vtk_to_numpy(self.GetStrips().GetData())
